+++ src/main/java/org/fugerit/java/query/export/catalog/QueryConfig.java
@@ -29,5 +29,7 @@ public class QueryConfig extends BasicIdConfigType {
 	@Getter @Setter private String xlsTemplate;
 	
 	@Getter @Setter private String createPath;
+
+	@Getter @Setter private String tryColumnType;
 	
 }