+++ src/main/java/org/fugerit/java/query/export/facade/format/QueryExportHandlerXLSBase.java
@@ -6,10 +6,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Iterator;
 
-import org.apache.poi.ss.usermodel.Cell;
-import org.apache.poi.ss.usermodel.Row;
-import org.apache.poi.ss.usermodel.Sheet;
-import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.ss.usermodel.*;
 import org.fugerit.java.core.function.SafeFunction;
 import org.fugerit.java.core.lang.helpers.BooleanUtils;
 import org.fugerit.java.query.export.facade.QueryExportConfig;
@@ -34,14 +31,29 @@ public abstract class QueryExportHandlerXLSBase extends QueryExportHandler {
 		super(format);
 	}
 	
-	private void addRow( Iterator<MetaField> currentRecord, Sheet sheet, int index ) {
+	private void addRow( Workbook workbook, Iterator<MetaField> currentRecord, Sheet sheet, int index, QueryExportConfig config ) {
 		int col = 0;
 		Row row = sheet.createRow( index );
 		while ( currentRecord.hasNext() ) {
 			MetaField field = currentRecord.next();
-			String value = field.getStringValue();
 			Cell cell = row.createCell( col );
-			cell.setCellValue( value );
+			if ( config.isTryColumnType() && field.getType() != MetaField.TYPE_STRING ) {
+				if ( field.getType() == MetaField.TYPE_DATE ) {
+					DataFormat format = workbook.createDataFormat();
+					CellStyle style = workbook.createCellStyle();
+					style.setDataFormat( format.getFormat( "m/d/yy h:mm" ) );
+					cell.setCellValue(  field.getTimestampValue() );
+					cell.setCellStyle(style);
+				} else if ( field.getType() == MetaField.TYPE_NUMBER ) {
+					DataFormat format = workbook.createDataFormat();
+					CellStyle style = workbook.createCellStyle();
+					style.setDataFormat( format.getFormat( "#,###" ) );
+					cell.setCellValue( field.getNumberValue() );
+					cell.setCellStyle(style);
+				}
+			} else {
+				cell.setCellValue( field.getStringValue() );
+			}
 			col++;
 		}
 	}
@@ -71,7 +83,7 @@ public abstract class QueryExportHandlerXLSBase extends QueryExportHandler {
 				if ( xlsTemplate == null ) {
 					sheet = workbook.createSheet();
 					if ( meta.hasHeader() ) {
-						addRow( meta.headerIterator() , sheet, 0 );
+						addRow( workbook, meta.headerIterator() , sheet, 0, config );
 					}
 				} else {
 					sheet = workbook.getSheetAt( 0 );
@@ -80,7 +92,7 @@ public abstract class QueryExportHandlerXLSBase extends QueryExportHandler {
 				Iterator<MetaRecord> itRec = meta.recordIterator();
 				while ( itRec.hasNext() ) {
 					MetaRecord currentRecord = itRec.next();
-					addRow( currentRecord.fieldIterator() , sheet, index );
+					addRow( workbook, currentRecord.fieldIterator() , sheet, index, config );
 					index++;
 				}
 				if ( BooleanUtils.isTrue( config.getParams().getProperty( QueryExportFacade.ARG_XLS_RESIZE ) ) ) {