+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- param tryColumnType, will try to format column by type.
+
 ## [1.0.1] - 2024-07-16
 
 ### Added